+++ src/plugins/stake-plugins/stakePlugins.ts
@@ -3,6 +3,7 @@ import { makeTronStakePlugin } from './currency/tronStakePlugin'
 import { makeGenericStakePlugin } from './generic/GenericStakePlugin'
 import { genericPlugins } from './generic/pluginInfo'
 import { makeTcSaversPlugin } from './thorchainSavers/tcSaversPlugin'
+import { makeTcSaversPluginSegwit } from './thorchainSavers/tcSaversPluginSegwit'
 import { makeThorchainTcyPlugin } from './thorchainSavers/tcyPlugin'
 import { StakePlugin } from './types'
 import { makeUniV2StakePlugin } from './uniswapV2/uniV2Plugin'
@@ -21,6 +22,9 @@ export const getStakePlugins = async (pluginId: string): Promise<StakePlugin[]> 
       console.warn(e.message)
     }),
     makeTcSaversPlugin(pluginId, { initOptions: tcInitOptions }).catch(e => {
+      console.warn(e.message)
+    }),
+    makeTcSaversPluginSegwit(pluginId, { initOptions: tcInitOptions }).catch(e => {
       console.warn(e.message)
     }),
     makeThorchainTcyPlugin(pluginId, { initOptions: tcInitOptions }).catch(e => {